fastApiMongo/requirements.txt
anyio==3.6.1
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
h11==0.13.0
httptools==0.4.0
idna==3.3
motor==3.0.0
pycodestyle==2.9.1
pydantic==1.9.2
PyJWT==2.4.0
pymongo==4.2.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3